refactor: Раскидал библиотечные вещи в основной код
